requirements.txt
allure-pytest==2.10.0
Appium-Python-Client==2.6.2
pydantic==1.10.2
pytest==7.1.3
selenium==4.4.3